requirements.txt
jira
playwright
openai>=1.0.0
python-dotenv